+++ src/main/java/coollaafi/wot/common/BaseEntity.java
@@ -8,22 +8,24 @@ import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import java.time.LocalDateTime;
 
+import jakarta.persistence.Column;
+
 @MappedSuperclass
 @EntityListeners(AuditingEntityListener.class)
 public abstract class BaseEntity {
 	@CreatedDate
-	private LocalDateTime createdAt;
-	@LastModifiedDate
-	private LocalDateTime updatedAt;
+	@Column(updatable = false, nullable = false)
+	private LocalDateTime created_at;
 
-	public BaseEntity() {
-	}
+	@LastModifiedDate
+	@Column(nullable = false)
+	private LocalDateTime updated_at;
 
 	public LocalDateTime getCreatedAt() {
-		return this.createdAt;
+		return created_at;
 	}
 
 	public LocalDateTime getUpdatedAt() {
-		return this.updatedAt;
+		return updated_at;
 	}
 }
