+++ src/main/java/coollaafi/wot/web/kakao/KakaoService.java
@@ -6,9 +6,9 @@ import coollaafi.wot.apiPayload.code.status.ErrorStatus;
 import coollaafi.wot.jwt.AuthTokens;
 import coollaafi.wot.jwt.AuthTokensGenerator;
 import coollaafi.wot.web.login.dto.LoginResponseDTO;
-import coollaafi.wot.member.handler.MemberHandler;
-import coollaafi.wot.member.repository.MemberRepository;
-import coollaafi.wot.member.entity.Member;
+import coollaafi.wot.web.member.handler.MemberHandler;
+import coollaafi.wot.web.member.repository.MemberRepository;
+import coollaafi.wot.web.member.entity.Member;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
