+++ src/main/java/coollaafi/wot/web/member/handler/MemberHandler.java
@@ -1,4 +1,4 @@
-package coollaafi.wot.member.handler;
+package coollaafi.wot.web.member.handler;
 
 import coollaafi.wot.apiPayload.code.BaseErrorCode;
 import coollaafi.wot.apiPayload.exception.GeneralException;
